+++ pkg/networkserver/grpc_gsns.go
@@ -953,9 +953,6 @@ func (ns *NetworkServer) handleDataUplink(ctx context.Context, up *ttnpb.UplinkM
 
 			stored = matched.Device
 			paths := ttnpb.AddFields(matched.SetPaths,
-				"mac_state.desired_parameters.adr_data_rate_index",
-				"mac_state.desired_parameters.adr_nb_trans",
-				"mac_state.desired_parameters.adr_tx_power_index",
 				"mac_state.recent_uplinks",
 			)
 			stored.MACState.RecentUplinks = appendRecentUplink(stored.MACState.RecentUplinks, &ttnpb.UplinkMessage{
@@ -967,18 +964,26 @@ func (ns *NetworkServer) handleDataUplink(ctx context.Context, up *ttnpb.UplinkM
 				DeviceChannelIndex: up.DeviceChannelIndex,
 				ConsumedAirtime:    up.ConsumedAirtime,
 			}, recentUplinkCount)
-			if !pld.FHDR.ADR {
+			useADR := mac.DeviceUseADR(stored, ns.defaultMACSettings, matched.phy)
+			if useADR {
+				if !pld.FHDR.ADR {
+					paths = ttnpb.AddFields(paths,
+						"mac_state.current_parameters.adr_data_rate_index",
+						"mac_state.current_parameters.adr_tx_power_index",
+					)
+					stored.MACState.CurrentParameters.ADRDataRateIndex = ttnpb.DATA_RATE_0
+					stored.MACState.CurrentParameters.ADRTxPowerIndex = 0
+				}
 				paths = ttnpb.AddFields(paths,
-					"mac_state.current_parameters.adr_data_rate_index",
-					"mac_state.current_parameters.adr_tx_power_index",
+					"mac_state.desired_parameters.adr_data_rate_index",
+					"mac_state.desired_parameters.adr_nb_trans",
+					"mac_state.desired_parameters.adr_tx_power_index",
 				)
-				stored.MACState.CurrentParameters.ADRDataRateIndex = ttnpb.DATA_RATE_0
-				stored.MACState.CurrentParameters.ADRTxPowerIndex = 0
+				stored.MACState.DesiredParameters.ADRDataRateIndex = stored.MACState.CurrentParameters.ADRDataRateIndex
+				stored.MACState.DesiredParameters.ADRTxPowerIndex = stored.MACState.CurrentParameters.ADRTxPowerIndex
+				stored.MACState.DesiredParameters.ADRNbTrans = stored.MACState.CurrentParameters.ADRNbTrans
 			}
-			stored.MACState.DesiredParameters.ADRDataRateIndex = stored.MACState.CurrentParameters.ADRDataRateIndex
-			stored.MACState.DesiredParameters.ADRTxPowerIndex = stored.MACState.CurrentParameters.ADRTxPowerIndex
-			stored.MACState.DesiredParameters.ADRNbTrans = stored.MACState.CurrentParameters.ADRNbTrans
-			if !pld.FHDR.ADR || !mac.DeviceUseADR(stored, ns.defaultMACSettings, matched.phy) {
+			if !pld.FHDR.ADR || !useADR {
 				return stored, paths, nil
 			}
 			if err := mac.AdaptDataRate(ctx, stored, matched.phy, ns.defaultMACSettings); err != nil {