+++ pkg/webui/console/views/gateway-general-settings/lorawan-settings-form/index.js
@@ -120,6 +120,7 @@ const LorawanSettingsForm = React.memo(props => {
       <Form.Field
         title={sharedMessages.scheduleAnyTimeDelay}
         name="schedule_anytime_delay"
+        inputWidth="s"
         component={UnitInput}
         description={{
           ...sharedMessages.scheduleAnyTimeDescription,